ASMO-1616 Prtg Errormessage is not visible in prtg because http status 503 if any check is failing

+++ src/Samhammer.AspNetCore.HealthChecks.Prtg/PrtgResponseWriter.cs
@@ -23,7 +23,7 @@ namespace Samhammer.AspNetCore.HealthChecks.Prtg
             var text = JsonConvert.SerializeObject(prtgRoot, SerializerSettings);
 
             httpContext.Response.ContentType = "application/json";
-            httpContext.Response.StatusCode = HttpStatusCode.OK;
+            httpContext.Response.StatusCode = (int)HttpStatusCode.OK;
             return httpContext.Response.WriteAsync(text);
         }
